requirements.txt
pyside6
python-dotenv
telethon
pylint
peewee
